refactor: put different middlewares under their own namespaces
